+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InlineQueries/InlineQueryResult/abstracts/results/video/InlineQueryResultVideoCommon.kt
@@ -1,6 +1,6 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult.abstracts.results.video
 
-import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.Captioned
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CaptionedOutput
 import com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult.abstracts.*
 
 const val inlineQueryResultVideoType = "video"
@@ -8,5 +8,5 @@ const val inlineQueryResultVideoType = "video"
 interface InlineQueryResultVideoCommon : InlineQueryResult,
     TitledInlineQueryResult,
     DescribedInlineQueryResult,
-    Captioned,
+    CaptionedOutput,
     WithInputMessageContentInlineQueryResult