+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/MediaGroupMemberInputMedia.kt
@@ -1,10 +1,11 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CaptionedOutput
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.StringFormat
 
 @Serializable(MediaGroupMemberInputMediaSerializer::class)
-interface MediaGroupMemberInputMedia: InputMedia {
+interface MediaGroupMemberInputMedia : InputMedia, CaptionedOutput {
     fun serialize(format: StringFormat): String
     val arguments: Map<String, Any?>
 }