+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/abstracts/CaptionedMediaContent.kt
@@ -1,8 +1,9 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts
 
-import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.MessageEntity
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CaptionedInput
 
-interface CaptionedMediaContent {
-    val caption: String?
-    val captionEntities: List<MessageEntity>
-}
+@Deprecated(
+    "Deprecated due to replace by common type CaptionedInput",
+    ReplaceWith("CaptionedInput", "com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CaptionedInput")
+)
+typealias CaptionedMediaContent = CaptionedInput