+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/abstracts/MediaGroupContent.kt
@@ -1,7 +1,8 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CaptionedInput
 import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.MediaGroupMemberInputMedia
 
-interface MediaGroupContent : MediaContent {
+interface MediaGroupContent : MediaContent, CaptionedInput {
     fun toMediaGroupMemberInputMedia(): MediaGroupMemberInputMedia
 }